Fix bug in dataset collection
Our dataset design assumes that there is a 100ms difference between any
two lines in the same sequence. However, I was not properly doing this
when the time_diff was zero. I should have skipped adding a line there,
but I did not.

Actually, the last commit already fixed that (accidentally). This diff
also makes it so in such cases, we overwrite the old cmd_vel with the
new one.

It is still very odd that two cmd_vels would be sent essentially
simultaneously. It may be worth investigating why this is happening by
digging into the original rosbag collection process.