Refactor: separate x from y in models.py, not train.py
This allows different models to decide for themselves which features are
independent vs. decision variables. From an ML standpoint this doesn't
make a ton of sense, but for instance it will allow us to write
baselines with oracle access to get some sense of the best possible
performance.

This is a refactor; it should have no effect on the output of the
program.